+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Release 0.1.12
 * Add convenience methods file, versions, and entry
 * **FIX**: Metadata had `contenttype` instead of `contentType`.
+* **FIX**: Fix a version-handling bug.
+* **FIX**: String decoding.
+* **FIX**: Metadata is merged.
 * **DOC**: Fix `README.md` links to take advantage of published current documentation.
 * **DOC**: Document the Continuous Integration.